sfcram3 goes public
Former-commit-id: 2e4a87c4a141002bd5e96b48cc4f72b758cd7e3d

+++ user/cram/cram_data2.c
@@ -135,7 +135,7 @@ float sf_cram_data2_get_sample (sf_cram_data2 cram_data, size_t i, float t,
     float *trace = NULL;
 
     if (cram_data->kmah)
-        i *= 2;
+        i *= (size_t)2;
 
     if (i >= cram_data->n)
         return 0.0;
